requirements.txt
fastapi
uvicorn[standard]
python-dotenv

langchain
langchain-community
langchain-google-genai

chromadb
sentence-transformers

pydantic
